+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Changed
 Fixed
 ^^^^^
 - Fix the total loss reported by the trainer (:gh:`515` by `Jérémy Scanvic`_)
+- Fix the gradient norm reported by the trainer (:gh:`520` by `Jérémy Scanvic`_)
 
 
 v0.3.1